+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.3.0](https://github.com/lightnow-ai/lightnow-proxy/compare/v1.2.0...v1.3.0) (2026-07-12)
+
+
+### Features
+
+* **#5:** resolve Vault bindings locally ([bcbb7b9](https://github.com/lightnow-ai/lightnow-proxy/commit/bcbb7b9e330fc73af861937b2e3ccc3a8fb01687))
+* report proxy health failure details ([d078020](https://github.com/lightnow-ai/lightnow-proxy/commit/d078020416abeecbf470f4ee41dd798fcf76c97b))
+* **secrets:** resolve Vault bindings locally ([1e78135](https://github.com/lightnow-ai/lightnow-proxy/commit/1e78135ecb1b889f76507eca878b30b5fa3fdcad))
+
+
+### Bug Fixes
+
+* default local proxy health config ([7e86405](https://github.com/lightnow-ai/lightnow-proxy/commit/7e8640557ff4e54f6eb09fe11dab84685c48ea0d))
+
 ## [1.2.0](https://github.com/lightnow-ai/lightnow-proxy/compare/v1.1.0...v1.2.0) (2026-07-06)